% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cv.R
\name{calc_cv}
\alias{calc_cv}
\alias{calc_cv.nutr_input}
\alias{calc_cv.meta_rn}
\title{calc_cv}
\usage{
calc_cv(x, what, timestep, verbose)

\method{calc_cv}{nutr_input}(x, what = NULL, timestep = NULL, verbose = TRUE)

\method{calc_cv}{meta_rn}(x, what = "ag_biomass", timestep = x$max_i, verbose = TRUE)
}
\arguments{
\item{x}{\code{nutr_input} or \code{meta_rn} object.}

\item{what}{String specifying which column us used for \code{meta_rn} object.}

\item{timestep}{Numeric with maximum timestep Only used for \code{meta_rn} object.}

\item{verbose}{Logical if TRUE progress reports are printed.}
}
\value{
list
}
\description{
Simulate nutrient input.
}
\details{
Calculates coefficient of variation on alpha, beta and gamma scale, where \emph{i}
refers to values on local scale and \emph{m} refers to the sum of all values
in the metaecosystem.

\deqn{\alpha_{i} = sd(x_{i}) / mean(x_{i})}

\deqn{\alpha = sum(mean(x_{i}) / mean(x_{m}) * \alpha_{i})}

\deqn{\gamma = sd(x_{m}) / mean(x_{m})}

\deqn{\beta = \alpha / \gamma}

\deqn{synchrony = var(x_{m}) / (sum(\alpha_{i}) ^ 2)}
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)
calc_cv(nutr_input)

\dontrun{
calc_cv(result)
}

}
\references{
Wang, S., Loreau, M., 2014. Ecosystem stability in space: α, β and γ
variability. Ecol Lett 17, 891–901. \url{https://doi.org/10.1111/ele.12292}

Wang, S., Loreau, M., 2016. Biodiversity and ecosystem stability across scales
in metacommunities. Ecol Lett 19, 510–518. \url{https://doi.org/10.1111/ele.12582}
}
