% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_variability.R
\name{calc_variability}
\alias{calc_variability}
\alias{calc_variability.nutr_input}
\alias{calc_variability.meta_rn}
\title{calc_variability}
\usage{
calc_variability(x, what, lag, verbose)

\method{calc_variability}{nutr_input}(x, what = NULL, lag = NULL, verbose = TRUE)

\method{calc_variability}{meta_rn}(x, what = "biomass", lag = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{\code{nutr_input} or \code{meta_rn} object.}

\item{what}{String specifying which measure to use to calculate variability.}

\item{lag}{Logical if TRUE, the difference to the previous timestep is returned.}

\item{verbose}{Logical if TRUE progress reports are printed.}
}
\value{
list
}
\description{
Calculate variability
}
\details{
Calculates coefficient of variation on alpha, beta and gamma scale, where \emph{i}
refers to values on local scale and \emph{m} refers to the sum of all values
in the metaecosystem.

The 'what' argument can be either 'biomass', 'production' or 'turnover'.

\deqn{\alpha_{i} = sd(x_{i}) / mean(x_{i})}

\deqn{\alpha = sum(mean(x_{i}) / mean(x_{m}) * \alpha_{i})}

\deqn{\gamma = sd(x_{m}) / mean(x_{m})}

\deqn{\beta = \alpha / \gamma}

\deqn{synchrony = var(x_{m}) / (sum(\alpha_{i}) ^ 2)}
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)
calc_variability(nutr_input)

\dontrun{
calc_variability(result)
}

}
\references{
Wang, S., Loreau, M., 2014. Ecosystem stability in space: α, β and γ
variability. Ecol Lett 17, 891–901. \url{https://doi.org/10.1111/ele.12292}

Wang, S., Loreau, M., 2016. Biodiversity and ecosystem stability across scales
in metacommunities. Ecol Lett 19, 510–518. \url{https://doi.org/10.1111/ele.12582}
}
