% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rstationary.R
\name{create_rstationary}
\alias{create_rstationary}
\title{create_rstationary}
\usage{
create_rstationary(fishpop_values, mean, sd)
}
\arguments{
\item{fishpop_values}{List with fishpopulation}

\item{mean, sd}{Numeric with mean and sd of log-norm distribution.}
}
\value{
matrix
}
\description{
Create random stationary values
}
\details{
Creates matrix with id and random stationary value for each individual of fishpopulation.
}
\examples{
# Add example code

}
