% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_starting}
\alias{default_starting}
\title{Default starting values}
\format{
List
}
\usage{
default_starting
}
\description{
List with some all starting values
}
\references{
ADD REFERENCES
}
\keyword{datasets}
