% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metarn.R
\name{filter_metarn}
\alias{filter_metarn}
\title{filter_metarn}
\usage{
filter_metarn(result, filter = c(0, result$max_i))
}
\arguments{
\item{result}{meta_rn object of simulation run.}

\item{filter}{Vector with timesteps to select}
}
\value{
meta_rn
}
\description{
Filter model run for specific timestep.
}
\details{
This functions allows to return only specific timesteps of a \code{meta_rn} object
created with \code{\link{run_meta}}. The function ensures that the object will
still be a \code{meta_rn} object.
}
\examples{
 \dontrun{
filter_metarn(result = result_rand, timestep = 10200)
}

}
