% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nutr_input.R
\name{filter_nutr_input}
\alias{filter_nutr_input}
\title{filter_nutr_input}
\usage{
filter_nutr_input(nutr_input, timesteps, verbose = TRUE)
}
\arguments{
\item{nutr_input}{\code{nutr_input} object.}

\item{timesteps}{Vector with timesteps to return.}

\item{verbose}{Logical if TRUE progress reports are printed.}
}
\value{
nutr_input
}
\description{
Filter nutrient input
}
\details{
Filter the nutrient input of each local metaecosystem to only include timesteps
specified by the corresponding function argument. This can be important, if during
the model run not all timesteps are returned in resulting object.
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)
filter_nutr_input(nutr_input, timesteps = seq(from = 1, to = 4380, by = 50))

}
