% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abundance.R
\name{get_abundance}
\alias{get_abundance}
\title{get_abundance}
\usage{
get_abundance(resulti)
}
\arguments{
\item{result}{meta_rn object of simulation run.}
}
\value{
data.frame
}
\description{
Get abundance of local metaecosystems
}
\details{
Get the number of individuals within each local metaecosystem.
}
\examples{
# Add example code

}
