% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_input.R
\name{get_global_input}
\alias{get_global_input}
\title{get_global_input}
\usage{
get_global_input(x, long = FALSE)
}
\arguments{
\item{x}{\code{nutr_input} object simulated with \code{sim_nutr_input}.}

\item{long}{Logical if TRUE, \code{data.frame} will be reshaped to long format.}
}
\value{
data.frame
}
\description{
Get global range
}
\details{
Returns a \code{data.frame} with all local nutrient inputs and total nutrients
input per timestep.
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)
get_global_input(nutr_input)

}
