% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_densities.R
\name{get_meta_densities}
\alias{get_meta_densities}
\title{get_meta_densities}
\usage{
get_meta_densities(
  result,
  timestep = result$max_i,
  normalize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{result}{List with \code{meta_rn} objects simulated with \code{simulate_nutr_input}.}

\item{timestep}{Numeric with time step to select}

\item{normalize}{Logical if TRUE count is divided by time steps.}

\item{verbose}{Logical if warning messages should be printed.}
}
\value{
data.frame
}
\description{
Get metaecosystems densities
}
\details{
Returns the densities (i.e., number of fish) within each cell.
}
\examples{
\dontrun{
get_meta_densities(result = result_attr)
}

}
