% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.arrR_parameters}
\alias{meta.arrR_parameters}
\title{Default parameters}
\format{
List
}
\usage{
meta.arrR_parameters
}
\description{
List with all default parameters.
}
\references{
ADD REFERENCES
}
\keyword{datasets}
