% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.arrR_starting_values}
\alias{meta.arrR_starting_values}
\title{Default starting values}
\format{
List
}
\usage{
meta.arrR_starting_values
}
\description{
List with some all starting values
}
\references{
ADD REFERENCES
}
\keyword{datasets}
