% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meta_rn.R
\name{plot.meta_rn}
\alias{plot.meta_rn}
\title{plot.meta_rn}
\usage{
\method{plot}{meta_rn}(x, what = "seafloor", burn_in = TRUE, base_size = 10, ...)
}
\arguments{
\item{x}{mdl_rn object of simulation run.}

\item{what}{Character specifying what to plot.}

\item{burn_in}{If TRUE, line to indicate burn-in time is plotted.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Ploting method for meta_rn object
}
\details{
Ploting method for metaecosystem created with \code{\link{setup_meta}}.
}
\examples{
# Add example code

}
