% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meta_syst.R
\name{plot.meta_syst}
\alias{plot.meta_syst}
\title{plot.meta_syst}
\usage{
\method{plot}{meta_syst}(x, ...)
}
\arguments{
\item{x}{\code{meta_syst} object simulated with \code{setup_meta}.}

\item{...}{Not used.}
}
\value{
ggplot
}
\description{
Plotting method for meta_syst object.
}
\details{
Plotting method for metaecosystem created with \code{setup_meta}.
}
\examples{
\dontrun{
plot(metasyst)
}

}
