% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fish_attr.R
\name{plot_fish_attr}
\alias{plot_fish_attr}
\title{plot_fish_attr}
\usage{
plot_fish_attr(x)
}
\arguments{
\item{x}{\code{meta_syst} object simulated with \code{setup_meta}.}
}
\value{
ggplot
}
\description{
Plot fish attributes
}
\details{
Plots the distribution of \code{reserves_thres} and \code{move_prob}.
}
\examples{
\dontrun{
plot_fish_attr(x = metasyst)
}

}
