% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_local_abund.R
\name{plot_local_abund}
\alias{plot_local_abund}
\title{plot_local_abund}
\usage{
plot_local_abund(x, base_size = 10, ...)
}
\arguments{
\item{x}{ntr_inpt object.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Ploting method for meta_rn object
}
\details{
Ploting method for nutrient iinput created with \code{\link{simulate_input_sine}}.
}
\examples{
# Add example code

}
