% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nutrient_input.R
\name{plot_nutrient_input}
\alias{plot_nutrient_input}
\title{plot_nutrient_input}
\usage{
plot_nutrient_input(x, base_size = 10, limits = NULL, ...)
}
\arguments{
\item{x}{ntr_inpt object.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Ploting method for ntr_inpt object
}
\details{
Ploting method for nutrient iinput created with \code{\link{simulate_input_sine}}.
}
\examples{
# Add example code

}
