% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_production.R
\name{plot_production}
\alias{plot_production}
\title{plot_production}
\usage{
plot_production(result, lag = TRUE, base_size = 10, ...)
}
\arguments{
\item{result}{\code{meta_rn} object simulated with \code{run_meta}.}

\item{lag}{Logical if TRUE, the difference to the previous timestep is returned.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Plot local abundance.
}
\details{
Plot the production per meta ecosystem
}
\examples{
\dontrun{
plot_production(result = result_rand)
}

}
