% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meta_rn.R
\name{print.meta_rn}
\alias{print.meta_rn}
\title{print.meta_rn}
\usage{
\method{print}{meta_rn}(x, digits = 2, timestep = x$max_i, ...)
}
\arguments{
\item{x}{meta_rn object.}

\item{digits}{Numeric of decimal places (round).}

\item{timestep}{Numeric with timestep to print.}

\item{...}{Arguments passed to cat.}
}
\description{
Printing method for meta_rn object
}
\details{
Printing method for metaecosystem model run created with \code{\link{simulate_meta}}.
}
\examples{
# Add example code

}
