% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meta_syst.R
\name{print.meta_syst}
\alias{print.meta_syst}
\alias{print.meta_sys}
\title{print.meta_syst}
\usage{
\method{print}{meta_syst}(x, digits = 3, ...)
}
\arguments{
\item{x}{mdl_rn object.}

\item{digits}{Numeric of decimal places (round).}

\item{...}{Arguments passed to cat.}
}
\description{
Printing method for mdl_rn object
}
\details{
Printing method for metaecosystem created with \code{\link{setup_meta}}.
}
\examples{
# Add example code

}
