% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nutr_input.R
\name{print.nutr_input}
\alias{print.nutr_input}
\title{print.nutr_input}
\usage{
\method{print}{nutr_input}(x, digits = NULL, ...)
}
\arguments{
\item{x}{\code{nutr_input} object simulated with \code{sim_nutr_input_*}.}

\item{digits}{Numeric of decimal places (passed on to \code{round}).}

\item{...}{Not used.}
}
\description{
Printing method for nutr_input object.
}
\details{
Printing method for \code{nutr_input} created with \code{sim_nutr_input_*}.
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)
print(nutr_input)

}
