% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_find}
\alias{rcpp_find}
\title{rcpp_find}
\usage{
rcpp_find(x, y)
}
\arguments{
\item{x}{Integer with with value to find.}

\item{y}{NumericVector with values to find all elements of x in.}
}
\value{
NumericVector
}
\description{
Rcpp find.
}
\details{
Returns the index of the \code{x} object within the \code{y} object.
The index of the first element is 0. If an element is not present within \code{y} object,
\code{NA} is returned. Returns only the first index of an object if present multiple
times in the \code{y} object.
}
