% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_table}
\alias{rcpp_get_table}
\title{rcpp_get_table}
\usage{
rcpp_get_table(x, n)
}
\arguments{
\item{x}{List}
}
\value{
vector
}
\description{
Rcpp get table
}
\details{
Returns table with value count
}
\keyword{export}
