% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_unique_values}
\alias{rcpp_get_unique_values}
\title{rcpp_get_unique_values}
\usage{
rcpp_get_unique_values(x)
}
\arguments{
\item{x}{vector}
}
\value{
matrix
}
\description{
Rcpp get unique values
}
\details{
Returns unique values from a vector
}
\keyword{export}
