% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_list_to_matrix}
\alias{rcpp_list_to_matrix}
\title{rcpp_list_to_matrix}
\usage{
rcpp_list_to_matrix(x, n, pop_n)
}
\arguments{
\item{x}{List}
}
\value{
matrix
}
\description{
Rcpp list to matrix
}
\details{
Convert a list to a list
}
\keyword{export}
