% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_matrix_to_list}
\alias{rcpp_matrix_to_list}
\title{rcpp_matrix_to_list}
\usage{
rcpp_matrix_to_list(x, n)
}
\arguments{
\item{x}{List}
}
\value{
matrix
}
\description{
Rcpp list to matrix
}
\details{
Convert a list to a list
}
\keyword{export}
