% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_meta}
\alias{rcpp_move_meta}
\title{rcpp_move_meta}
\usage{
rcpp_move_meta(fishpop_values, n, pop_n, prob_move)
}
\arguments{
\item{x}{List}
}
\value{
matrix
}
\description{
Rcpp list to matrix
}
\details{
Convert a list to a list
}
\keyword{export}
