% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sim_processes}
\alias{rcpp_sim_processes}
\title{rcpp_sim_processes}
\usage{
rcpp_sim_processes(
  seafloor,
  fishpop,
  seafloor_track,
  fishpop_track,
  parameters,
  movement,
  max_dist,
  n,
  pop_n,
  fishpop_attributes,
  nutr_input,
  coords_reef,
  cell_adj,
  extent,
  dimensions,
  max_i,
  min_per_i,
  save_each,
  seagrass_each,
  burn_in,
  verbose
)
}
\arguments{
\item{seafloor, fishpop}{List with seafloor and fishpop data of metaecosystems.}

\item{seafloor_track, fishpop_track}{List with entry for each saving timestep and metaecosystems.}

\item{parameters}{List with parameters.}

\item{movement}{String specifying movement algorithm. Either 'rand', 'attr' or 'behav'.}

\item{max_dist}{Double with maximum movement distance.}

\item{n}{Integer with number of metaecosystems.}

\item{pop_n}{Vector with number of individuals.}

\item{fishpop_attributes}{Matrix with stationary and reserves_thres values for each individual}

\item{nutr_input}{List with amount of nutrient input each timestep.}

\item{coords_reef}{List with ID and coords of reef cells.}

\item{cell_adj}{Matrix with cell adjacencies.}

\item{extent}{Vector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{Vector with dimensions (nrow, ncol).}

\item{max_i}{Integer with maximum number of simulation timesteps.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{save_each}{Numeric how often data should be saved to return.}

\item{seagrass_each}{Integer how often (each i * x) seagrass dynamics will be simulated.}

\item{burn_in}{Numeric with timesteps used to burn in.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
void
}
\description{
Rcpp run simulation of metaecosystems processes.
}
\details{
The functions is a 'wrapper' around the following sub-processes: (i) nutrient input,
(ii) seagrass growth, (iii) detritus mineralization, (iv) movement of individuals,
(v) respiration of individuals, (vi) growth of individuals, (vii) mortality of individuals,
(viii) diffusion of nutrients/detritus, and ix) nutrient output.
}
\references{
For a detailed model description, see Esquivel, K., Hesselbarth, M.H.K., Allgeier, J.E.
In preparation. Mechanistic support for increased primary production around artificial reefs.

Add references about meta approach.
}
