% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_meta}
\alias{rcpp_simulate_meta}
\title{rcpp_simulate_meta}
\usage{
rcpp_simulate_meta(
  seafloor,
  fishpop,
  nutrients_input,
  fishpop_attr,
  seafloor_probs,
  seafloor_track,
  fishpop_track,
  parameters,
  movement,
  extent,
  dimensions,
  max_i,
  min_per_i,
  save_each,
  seagrass_each,
  burn_in,
  verbose
)
}
\arguments{
\item{seafloor, fishpop}{List with seafloor and fishpop data of metaecosystems.}

\item{nutrients_input}{List with amount of nutrient input each timestep.}

\item{fishpop_attr}{NumericMatrix with residence and reserves_thres values for each individual}

\item{seafloor_probs}{NumericMatrix with local ecosystems probabilities.}

\item{seafloor_track, fishpop_track}{List with entry for each saving timestep and metaecosystems.}

\item{parameters}{List with parameters.}

\item{movement}{String specifying movement algorithm. Either 'rand', 'attr' or 'behav'.}

\item{extent}{NumericVector with extent (xmin,xmax,ymin,ymax).}

\item{dimensions}{IntegerVector with dimensions (nrow, ncol).}

\item{max_i}{Integer with maximum number of simulation timesteps.}

\item{min_per_i}{Integer to specify minutes per i.}

\item{save_each}{Numeric how often data should be saved to return.}

\item{seagrass_each}{Integer how often (each i * x) seagrass dynamics will be simulated.}

\item{burn_in}{Numeric with timesteps used to burn in.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
void
}
\description{
Rcpp run simulation of metaecosystems processes.
}
\details{
Wrapper function including all sub-processes. The core of the function is a
nested loop through all i) time steps and ii) all local metaecosystems.

The functions is a 'wrapper' around the following sub-processes: i) movement across
local metaecosystems (outter loop) (ii) nutrient input, (iii) seagrass growth,
(iv) detritus mineralization, (v) movement of individuals, (vi) respiration of
individuals, (vii) growth of individuals, (viii) mortality of individuals,
(ix) diffusion of nutrients/detritus, and (x) nutrient output (all inner loop).

For a detailed description of all sub-processes (with exception of (i)), see the
\pkg{arrR} package.
}
\references{
Add references
}
\keyword{internal}
