% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_subset_matrix}
\alias{rcpp_subset_matrix}
\title{rcpp_subset_matrix}
\usage{
rcpp_subset_matrix(fishpop, rows)
}
\arguments{
\item{fishpop}{Matrix with fishpop values.}

\item{rows}{Vector with row ids}
}
\value{
matrix
}
\description{
Rcpp subset matrix.
}
\details{
Returns matrix with only subset of rows of \code{fishpop} specified by \code{rows}.
}
\keyword{export}
