% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_update_behavior}
\alias{rcpp_update_behavior}
\title{rcpp_update_behavior}
\usage{
rcpp_update_behavior(fishpop, fishpop_behavior)
}
\arguments{
\item{fishpop}{NumericMatrix with fish population.}

\item{fishpop_behavior}{NumericMatrix with fishpop behavior}
}
\value{
void
}
\description{
Rcpp update behavior.
}
\details{
Updates behavior column in fishpop behavior matrix.
}
\keyword{internal}
