% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_which}
\alias{rcpp_which}
\title{rcpp_which}
\usage{
rcpp_which(x, y)
}
\arguments{
\item{x}{Vector with values.}

\item{y}{Vector values to find position.}
}
\value{
vector
}
\description{
Rcpp which.
}
\details{
Returns index of all elements of \code{y} vector within \code{x} vector.
The index of the first element is 0. If element is not present within  \code{x} vector
\code{NA} is returned.
}
\keyword{export}
