% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_variability.R
\name{sample_variability}
\alias{sample_variability}
\alias{sample_variability.nutr_input}
\alias{sample_variability.meta_rn}
\title{sample_variability}
\usage{
sample_variability(x, itr, biomass, production, lag, verbose)

\method{sample_variability}{nutr_input}(
  x,
  itr = 1,
  biomass = NULL,
  production = NULL,
  lag = NULL,
  verbose = TRUE
)

\method{sample_variability}{meta_rn}(
  x,
  itr = 1,
  biomass = TRUE,
  production = TRUE,
  lag = c(FALSE, FALSE),
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{nutr_input} or \code{meta_rn} object.}

\item{itr}{Integer with number of sample iterations.}

\item{biomass, production}{Logical to specifiy if biomass and/or production is summarize.d}

\item{lag}{Vector with logical. If TRUE, the difference to the previous timestep is returned.
The first element refers to biomass, the second element to production.}

\item{verbose}{Logical if TRUE, progress reports are printed.}
}
\value{
data.frame
}
\description{
Sample variability
}
\details{
Samples coefficient of variation (on gamma scale) for increasing number of metaecosystems. For
more information see \code{calc_variability}. The \code{itr} arguments allows to
resample the data n times to capture differences based on different orders in which
the metaecosystems are sampled.
}
\examples{
nutrients_input <- simulate_nutrient_sine(n = 3, max_i = 4380, input_mn = 1,
frequency = 3, noise = 0.5)
sample_variability(nutrients_input)

\dontrun{
sample_variability(result)
}

}
