% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_attributes.R
\name{setup_attributes}
\alias{setup_attributes}
\title{setup_attributes}
\usage{
setup_attributes(fishpop, parameters, max_i)
}
\arguments{
\item{fishpop}{List with fish population}

\item{parameters}{List with parameters}

\item{max_i}{Integer with maximum number of simulation time steps.}
}
\value{
matrix
}
\description{
Create attribute values.
}
\details{
Creates matrix with id, residence, and reserves threshold value for each
individual of fish population.
}
\examples{
\dontrun{
setup_attributes(fishpop = fishpop_list, parameters = parameters, max_i = 4380)
}

}
\keyword{internal}
