% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_meta.R
\name{setup_meta}
\alias{setup_meta}
\title{setup_meta}
\usage{
setup_meta(
  n,
  max_i,
  dimensions,
  grain = c(1, 1),
  reef = NULL,
  seafloor_xy = NULL,
  starting_values,
  parameters,
  random = 0,
  use_log = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{dimensions}{Vector with number of rows and columns (spatial dimensions).}

\item{grain}{Vector with size of cells in x- and y-direction (spatial grain).}

\item{reef}{2-Column matrix with coordinates of artificial reefs.}

\item{seafloor_xy}{3-Column matrix with ID and x,y coordinates of local ecosystems.}

\item{starting_values}{List with all starting value.}

\item{parameters}{List with all parameters.}

\item{random}{Numeric to randomize input values by 0 = 0 percent to 1 = 100 percent.}

\item{use_log}{Logical if TRUE, random log distribution is used}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{rast}.}
}
\value{
meta_syst
}
\description{
Setup metaecosystems.
}
\details{
Function to create metaecosystem including environmental raster grids and fish
populations.

If \code{seafloor_xy = NULL}, coordinates are simulated random (range: 0-1).
}
\examples{
metasyst <- setup_meta(n = 3, max_i = 4380, dimensions = c(100, 100), grain = c(1, 1),
starting_values = meta.arrR_starting_values, parameters = meta.arrR_parameters)

}
