% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nutr_input_sd.R
\name{sim_nutr_input_sd}
\alias{sim_nutr_input_sd}
\title{sim_nutr_input_sd}
\usage{
sim_nutr_input_sd(n, max_i, freq_mn, input_max, variability)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{freq_mn}{Numeric number of peak.}

\item{input_max}{Numeric with maximum input amount.}

\item{variability}{Variability of nutrient input}

\item{phase}{Numeric with sine curve parameters.}
}
\value{
vector
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curve. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_max} argument quantifies the
maximum value of the input "peaks". Both arguments can be multiplied by the \code{*_sd}
argument and used as standard deviation of a normal distribution to add variability across
metaecosystems.
}
\examples{
nutr_input <- sim_nutr_input_sd(n = 3, max_i = 4380, freq_mn = 3, freq_sd = 0.1,
input_max = 1, input_sd = 0.5)

}
