% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement_meta.R
\name{simulate_movement_meta}
\alias{simulate_movement_meta}
\title{simulate_movement_meta}
\usage{
simulate_movement_meta(fishpop_values)
}
\arguments{
\item{fishpop_values}{List with fishpop matrices.}
}
\value{
vector
}
\description{
Initiate fish population.
}
\details{
ADD TEXT
}
\examples{
# Add example code

}
