% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nutr_input.R
\name{simulate_nutr_input}
\alias{simulate_nutr_input}
\title{simulate_nutr_input}
\usage{
simulate_nutr_input(
  n,
  max_i,
  freq_mn,
  freq_var,
  input_max,
  input_var,
  phase = 0
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{freq_mn, freq_var}{Numeric number of peaks.}

\item{input_max, input_var}{Numeric with maximum input amount.}

\item{phase}{Numeric with sine curve parameters.}
}
\value{
vector
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curve. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_max} argument quantifies the
maximum value of the input "peaks". Both arguments can be multiplied by the \code{*_var}
argument and used as standard deviation of a normal distribution to add variability across
metaecosystems.
}
\examples{
\dontrun{
nutr_input <- simulate_nutr_input(n = n, max_i = max_i, freq_mn = 3, freq_var = 0.1,
input_max = input_max, input_var = 0.5)
}

}
