% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nutr_input_jitter.R
\name{simulate_nutr_input_jitter}
\alias{simulate_nutr_input_jitter}
\title{simulate_nutr_input_jitter}
\usage{
simulate_nutr_input_jitter(n, max_i, freq_mn, input_max, variability)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{freq_mn}{Numeric number of peaks.}

\item{input_max}{Numeric with maximum input amount.}

\item{variability}{Variability of nutrient input}
}
\value{
vector
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curve. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_max} argument quantifies the
maximum value of the input "peaks". Both arguments can be multiplied by the \code{*_var}
argument and used as standard deviation of a normal distribution to add variability across
metaecosystems.
}
\examples{
nutr_input <- simulate_nutr_input_jitter(n = 3, max_i = 4380, freq_mn = 3,
input_max = 1, variability = 0.5)

}
