% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nutr_input_noise.R
\name{simulate_nutr_input_noise}
\alias{simulate_nutr_input_noise}
\title{simulate_nutr_input_noise}
\usage{
simulate_nutr_input_noise(
  n,
  max_i,
  freq_mn,
  input_max,
  variability,
  n_noise = 3
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{freq_mn}{Numeric number of peaks.}

\item{input_max}{Numeric with maximum input amount.}

\item{variability}{Variability of nutrient input}

\item{n_noise}{Number of noise sine functions}
}
\value{
vector
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curve. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_max} argument quantifies the
maximum value of the input "peaks". Both arguments can be multiplied by the \code{variability}
argument and used as standard deviation of a normal distribution to add variability across
metaecosystems. Variability is added by taking the mean of \code{n_noise} sine curves.
}
\examples{
nutr_input <- simulate_nutr_input_noise(n = 3, max_i = 4380, freq_mn = 3,
input_max = 1, variability = 0.5)

}
