% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nutrient_sine.R
\name{simulate_nutrient_sine}
\alias{simulate_nutrient_sine}
\title{simulate_nutrient_sine}
\usage{
simulate_nutrient_sine(
  n,
  max_i,
  frequency = 0,
  input_mn = 0,
  noise = 0,
  verbose = TRUE
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{frequency}{Numeric frequency.}

\item{input_mn}{Numeric with mean input amount and frequency.}

\item{noise}{Numeric with noise function sd.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
nutr_input
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curves. The \code{frequency} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_mn} argument quantifies the
mean input values. Amplitude and phase mean and sd can be controlled by the corresponding
arguments.
}
\examples{
nutrients_input <- simulate_nutrient_sine(n = 3, max_i = 4380, frequency = 5,
input_mn = 1, noise = 0.5)
plot(nutrients_input)

}
