% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_probability.R
\name{calc_probability}
\alias{calc_probability}
\title{calc_probability}
\usage{
calc_probability(metasyst, lambda = 1)
}
\arguments{
\item{metasyst}{\code{meta_syst} object simulated with \code{setup_meta}.}

\item{lambda}{Distance decay parameter.}
}
\value{
matrix
}
\description{
Calculate probability
}
\details{
Calculate probability matrix for local ecosystems based on distances between them
using a negative exponential function I = exp(-d*lambda).
}
\examples{
\dontrun{
calc_probability(metasyst)
}

}
\references{
https://en.wikipedia.org/wiki/Exponential_decay

https://en.wikipedia.org/wiki/Distance_decay

Nekola, J.C. and White, P.S., 1999. The distance decay of similarity in biogeography
and ecology. Journal of biogeography, 26(4), pp.867-878.
}
