% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_probability.R
\name{calc_probability}
\alias{calc_probability}
\title{calc_probability}
\usage{
calc_probability(metasyst, lambda = 1)
}
\arguments{
\item{metasyst}{ADD TEXT}

\item{lambda}{Distance decay parameter.}
}
\value{
matrix
}
\description{
Calculate probability
}
\details{
ADD TEXT
}
\examples{
\dontrun{
calc_probability(metasyst)
}

}
