% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes.R
\name{create_attributes}
\alias{create_attributes}
\title{create_attributes}
\usage{
create_attributes(fishpop, parameters)
}
\arguments{
\item{fishpop}{List with fish population}

\item{parameters}{List with parameters}
}
\value{
matrix
}
\description{
Create attribute values.
}
\details{
Creates matrix with id, stationary, and reserves threshold value for each
individual of fish population.
}
\examples{
\dontrun{
create_attributes(fishpop = fishpop_list, parameters = parameters)
}

}
