% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_meta.R
\name{filter_meta}
\alias{filter_meta}
\alias{filter_meta.nutr_input}
\alias{filter_meta.meta_rn}
\title{filter_meta}
\usage{
filter_meta(x, filter, reset, verbose)

\method{filter_meta}{nutr_input}(x, filter, reset = NULL, verbose = TRUE)

\method{filter_meta}{meta_rn}(x, filter, reset = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{\code{nutr_input} or \code{meta_rn} object.}

\item{filter}{Integer with one or vector with min/max timestep(s) to filter.}

\item{reset}{Logical if TRUE, cumulative seafloor values are reduced by value
before filter minimum.}

\item{verbose}{Logical if TRUE progress reports are printed.}
}
\value{
nutr_input, meta_rn
}
\description{
Filter meta objects
}
\details{
Filter the nutrient input of each local metaecosystem to only include timesteps
specified by the corresponding function argument. This can be important, if during
the model run not all timesteps are returned in resulting object.
}
\examples{
nutrients_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, frequency = 3,
amplitude_sd = 0.5)
filter_meta(x = nutrients_input, filter = c(4380 / 2, 4380))

\dontrun{
filter_meta(x = result_rand, filter = c(4380 / 2, 4380))
}

}
