% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abundance.R
\name{get_abundance}
\alias{get_abundance}
\title{get_abundance}
\usage{
get_abundance(result)
}
\arguments{
\item{result}{\code{meta_rn} object simulated with \code{simulate_nutr_input}.}
}
\value{
data.frame
}
\description{
Get abundance of local metaecosystems.
}
\details{
Get the number of individuals within each local metaecosystem of model run result
created with \code{run_meta}.
}
\examples{
\dontrun{
get_abundance(result = result_attr)
}

}
