% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_df.R
\name{get_input_df}
\alias{get_input_df}
\title{get_input_df}
\usage{
get_input_df(x, long = FALSE)
}
\arguments{
\item{x}{\code{nutr_input} object simulated with \code{sim_nutr_input}.}

\item{long}{Logical if TRUE, \code{data.frame} will be reshaped to long format.}
}
\value{
data.frame
}
\description{
Get input data.frame
}
\details{
Returns a \code{data.frame} with all local nutrient inputs and total nutrients
input per timestep.
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)
get_input_df(nutr_input)

}
