% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.arrR-packages.R
\docType{package}
\name{meta.arrR}
\alias{meta.arrR}
\alias{meta.arrR-package}
\title{meta.arrR: Metaecosystems of Artifical Reefs}
\description{
Simulate metaecosystem of artifical reefs. For more information about how to
use the model, please see \code{browseVignettes("meta.arrR")}.
}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Other contributors:
\itemize{
  \item Jacob E. Allgeier \email{jeallg@umich.edu} (\href{https://orcid.org/0000-0002-9005-6432}{ORCID}) [research team head]
}

}
\keyword{internal}
