% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meta_rn.R
\name{plot.meta_rn}
\alias{plot.meta_rn}
\title{plot.meta_rn}
\usage{
\method{plot}{meta_rn}(x, what = "seafloor", burn_in = TRUE, limits = NULL, base_size = 10, ...)
}
\arguments{
\item{x}{meta_rn object.}

\item{...}{Arguments passed to cat.}

\item{digits}{Numeric of decimal places (round).}
}
\description{
Ploting method for meta_rn object
}
\details{
Ploting method for metaecosystem created with \code{\link{setup_meta}}.
}
\examples{
# Add example code

}
