% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meta_rn.R
\name{plot.meta_rn}
\alias{plot.meta_rn}
\title{plot.meta_rn}
\usage{
\method{plot}{meta_rn}(
  x,
  what = "seafloor",
  summarize = FALSE,
  fill = "ag_biomass",
  gamma = FALSE,
  limits = NULL,
  burn_in = FALSE,
  base_size = 10,
  viridis_option = "C",
  ...
)
}
\arguments{
\item{x}{\code{meta_rn} object simulated with \code{run_meta}.}

\item{what}{Character specifying what to plot.}

\item{summarize}{Logical if TRUE values over time steps are plotted.}

\item{fill}{Character specifying which column to use for plotting.}

\item{gamma}{Logical if TRUE gamma input line is added.}

\item{limits}{Vector with minium and maximum value of \code{fill} values.}

\item{burn_in}{If TRUE, line to indicate burn-in time is plotted.}

\item{base_size}{Numeric to specify base font size.}

\item{viridis_option}{Character with \code{viridis} color palette.}

\item{...}{Not used.}
}
\description{
Plotting method for meta_rn object.
}
\details{
Plotting method for result of metaecosystem model run created
with \code{run_meta}.
}
\examples{
\dontrun{
plot(result_rand)
}

}
