% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nutr_input.R
\name{plot.nutr_input}
\alias{plot.nutr_input}
\title{plot.nutr_input}
\usage{
\method{plot}{nutr_input}(x, what = "seafloor", burn_in = TRUE, base_size = 10, ...)
}
\arguments{
\item{x}{mdl_rn object of simulation run.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Ploting method for nutr_input object.
}
\details{
Ploting method for metaecosystem created with \code{\link{setup_meta}}.
}
\examples{
# Add example code

}
