% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nutr_input.R
\name{plot.nutr_input}
\alias{plot.nutr_input}
\title{plot.nutr_input}
\usage{
\method{plot}{nutr_input}(x, base_size = 10, ...)
}
\arguments{
\item{x}{\code{nutr_input} object simulated with \code{sim_nutr_input_*}.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Plotting method for nutr_input object.
}
\details{
Ploting method for metaecosystem created with \code{setup_meta}.
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, freq_mn = 3,
input_max = 1, variability = 0.5)
plot(nutr_input)

}
