% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_local_abund.R
\name{plot_local_abund}
\alias{plot_local_abund}
\title{plot_local_abund}
\usage{
plot_local_abund(x, base_size = 10, ...)
}
\arguments{
\item{x}{\code{meta_rn} object simulated with \code{run_meta}.}

\item{base_size}{Numeric to specify base font size.}

\item{...}{Not used.}
}
\description{
Plot local abundance.
}
\details{
The resulting plot describes the number of individuals within each local metaecosystem
over time steps. Additional, the minimum, maximum and mean of all metaecosystems are
included.
}
\examples{
\dontrun{
plot_local_abund(x = result_attr)
}

}
