% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meta_rn.R
\name{print.meta_rn}
\alias{print.meta_rn}
\title{print.meta_rn}
\usage{
\method{print}{meta_rn}(x, digits = 2, timestep = x$max_i, ...)
}
\arguments{
\item{x}{\code{meta_rn} object simulated with \code{run_meta}.}

\item{digits}{Numeric of decimal places (passed on to \code{round}).}

\item{timestep}{Numeric with timestep to print.}

\item{...}{Not used.}
}
\description{
Printing method for meta_rn object.
}
\details{
Printing method for metaecosystem model run created with \code{run_meta}.
Returns the mean values for several seafloor and fish population values.
}
\examples{
\dontrun{
print(result_rand)
}

}
