% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meta_syst.R
\name{print.meta_syst}
\alias{print.meta_syst}
\alias{print.meta_sys}
\title{print.meta_syst}
\usage{
\method{print}{meta_syst}(x, digits = 3, ...)
}
\arguments{
\item{x}{\code{meta_syst} object created with \code{setup_meta}.}

\item{digits}{Numeric of decimal places (passed on to \code{round}).}

\item{...}{Not used.}
}
\description{
Printing method for meta_syst object.
}
\details{
Printing method for metaecosystem created with \code{setup_meta}.
}
\examples{
nutr_input <- simulate_nutr_input(n = 3, max_i = 4380, freq_mn = 3, freq_var = 0.1,
input_max = 0.1, input_var = 0.25)

}
