% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nutr_input.R
\name{print.nutr_input}
\alias{print.nutr_input}
\title{print.nutr_input}
\usage{
\method{print}{nutr_input}(x, digits = NULL, ...)
}
\arguments{
\item{x}{\code{nutr_input} object simulated with \code{simulate_nutrient_*}.}

\item{digits}{Numeric of decimal places (passed on to \code{round}).}

\item{...}{Not used.}
}
\description{
Printing method for nutr_input object.
}
\details{
Printing method for \code{nutr_input} created with \code{simulate_nutr_input_*}.
}
\examples{
nutrients_input <- simulate_nutrient_sine(n = 3, max_i = 4380, input_mn = 1,
frequency = 3, noise = 0.5)
print(nutrients_input)

}
