% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nutr_input.R
\name{print.nutr_input}
\alias{print.nutr_input}
\title{print.nutr_input}
\usage{
\method{print}{nutr_input}(x, digits = NULL, ...)
}
\arguments{
\item{x}{nutr_input object.}

\item{digits}{Numeric of decimal places (round).}

\item{...}{Arguments passed to cat.}
}
\description{
Printing method for nutr_input object.
}
\details{
Printing method for nutr_input model run created with \code{simulate_input_sine}.
}
\examples{
# Add example code

}
