% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_find}
\alias{rcpp_find}
\title{rcpp_find}
\usage{
rcpp_find(x, y)
}
\arguments{
\item{x}{Integer with with value to find.}

\item{y}{NumericVector with values to find all elements of x in.}
}
\value{
int
}
\description{
Rcpp find.
}
\details{
Returns the index (starting at zero) of the \code{x} integer within the \code{y} vector.
If an element is not present within \code{y} vector, \code{NA} is returned.
Returns only the first index if \code{x} is present multiple times in the \code{y} vector.
}
