% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_list_to_matrix}
\alias{rcpp_list_to_matrix}
\title{rcpp_list_to_matrix}
\usage{
rcpp_list_to_matrix(x, pop_n)
}
\arguments{
\item{x}{List with individuals within metaecosystems.}

\item{pop_n}{Integer with number of indiviuals.}
}
\value{
matrix
}
\description{
Rcpp list to matrix
}
\details{
Converts list with individuals within metaecosystems to one matrix. Adds coloumn
with metasystem id
}
\keyword{export}
