% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_list_to_matrix}
\alias{rcpp_list_to_matrix}
\title{rcpp_list_to_matrix}
\usage{
rcpp_list_to_matrix(fishpop, pop_n_sum, id)
}
\arguments{
\item{fishpop}{List with individuals within local metaecosystems.}

\item{pop_n_sum}{Integer with number of individuals.}

\item{id}{If TRUE, column with metaecosystem id is added.}
}
\value{
NumericMatrix
}
\description{
Rcpp list to matrix.
}
\details{
Converts list with individuals within a local metaecosystem to one matrix. If
\code{id = TRUE} a column identifying the metaecosystem is added.
}
