% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_matrix_to_list}
\alias{rcpp_matrix_to_list}
\title{rcpp_matrix_to_list}
\usage{
rcpp_matrix_to_list(fishpop, n)
}
\arguments{
\item{fishpop}{Matrix with fish population.}

\item{n}{Integer with number of metaecosystems.}
}
\value{
list
}
\description{
Rcpp matrix to list
}
\details{
Converts a matrix with fishpopulation to a list split by metaecosystem.
}
\keyword{export}
