% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_meta}
\alias{rcpp_move_meta}
\title{rcpp_move_meta}
\usage{
rcpp_move_meta(
  seafloor_probs,
  fishpop,
  residence_values,
  n,
  pop_n_sum,
  id_attr,
  id_meta,
  extent
)
}
\arguments{
\item{seafloor_probs}{Matrix with local ecosystems probabilities.}

\item{fishpop}{List with fish population.}

\item{residence_values}{Vector with residence values.}

\item{n, pop_n_sum}{Integer with total number of local metaecosystems and individuals.}

\item{id_attr}{Vector with unique id of fishpop attributes matrix.}

\item{id_meta}{Vector with metaecosystem ids.}

\item{extent}{Spatial extent of the seafloor raster.}
}
\value{
list
}
\description{
Rcpp move meta.
}
\details{
Simulate movement across local metaecosystem. Individuals move to a new local
metaecosystem with a certain probability each timestep. The probability increases
depending on the residence value and how long individuals already stayed on local
metaecosystem. To avoid this movement set \code{parameters$move_residence = 0}.
}
