% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_meta}
\alias{rcpp_move_meta}
\title{rcpp_move_meta}
\usage{
rcpp_move_meta(fishpop, seafloor_probs, fishpop_attributes, extent)
}
\arguments{
\item{fishpop}{List with fish population.}

\item{seafloor_probs}{NumericMatrix with local ecosystems probabilities.}

\item{fishpop_attributes}{NumericMatrix with residence and reserves_thres values for each individual.}

\item{extent}{NumericVector with spatial extent of the seafloor raster.}
}
\value{
list
}
\description{
Rcpp move meta.
}
\details{
Simulate movement across local metaecosystem. Individuals move to a new local
metaecosystem with a certain probability each timestep. The probability increases
depending on the residence value and how long individuals already stayed on local
metaecosystem. To avoid this movement set \code{parameters$move_residence <= 0}.
}
