% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_meta}
\alias{rcpp_move_meta}
\title{rcpp_move_meta}
\usage{
rcpp_move_meta(fishpop, fishpop_attr, seafloor_probs, extent)
}
\arguments{
\item{fishpop}{List with fish population.}

\item{fishpop_attr}{NumericMatrix with reserves_thres and prob_move values for each individual.}

\item{seafloor_probs}{NumericMatrix with local ecosystems probabilities.}

\item{extent}{NumericVector with spatial extent of the seafloor raster.}
}
\value{
list
}
\description{
Rcpp move meta.
}
\details{
Simulate movement across local metaecosystem. Individuals move if their residence
counter equals the maximum residence time specified for each individual in the
attributes table. To avoid this movement set \code{parameters$move_meta_each = 0}.
}
\keyword{internal}
