% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_meta}
\alias{rcpp_move_meta}
\title{rcpp_move_meta}
\usage{
rcpp_move_meta(fishpop, pop_n_sum, id_attr, stationary_values, id_meta, extent)
}
\arguments{
\item{fishpop}{List with fish population.}

\item{pop_n_sum}{Integer with total number of individuals.}

\item{id_attr}{Vector with unique id of fishpop attributes matrix.}

\item{stationary_values}{Vector with stationary values.}

\item{id_meta}{Vector with metaecosystem ids.}

\item{extent}{Spatial extent of the seafloor raster.}
}
\value{
list
}
\description{
Rcpp move meta.
}
\details{
Simulate movement across local metaecosystem. Individuals move to a new local
metaecosystem with a certain probability each timestep. The probability increases
depending on the stationary value and how long individuals already stayed on local
metaecosystem. To avoid this movement set \code{parameters$move_stationary = 0}.
}
\keyword{export}
