% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_move_meta}
\alias{rcpp_move_meta}
\title{rcpp_move_meta}
\usage{
rcpp_move_meta(fishpop, pop_n_sum, id_attr, stationary_values, id_meta, extent)
}
\arguments{
\item{fishpop}{List with fish population.}

\item{pop_n_sum}{Integer with total number of individuals.}

\item{extent}{Spatial extent of the seafloor raster.}

\item{fishpop_attributes}{Matrix with stationary value for each individual.}
}
\value{
void
}
\description{
Rcpp list to matrix
}
\details{
Simulate movement across local metaecosystem. Individuals move to a new local
metaecosystem with a certain probability each timestep.
}
\keyword{export}
