% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_which}
\alias{rcpp_which}
\title{rcpp_which}
\usage{
rcpp_which(x, y)
}
\arguments{
\item{x}{Vector with values.}

\item{y}{Integer with maximum number of classes.}
}
\value{
vector
}
\description{
Rcpp which
}
\details{
ADD DETAILS
}
\keyword{export}
