% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_meta.R
\name{setup_meta}
\alias{setup_meta}
\title{setup_meta}
\usage{
setup_meta(
  n,
  extent,
  grain,
  reefs = NULL,
  starting_values,
  parameters,
  random = 0,
  use_log = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{extent}{Vector with number of rows and columns (spatial extent).}

\item{grain}{Vector with size of cells in x- and y-direction (spatial grain).}

\item{reefs}{2-Column matrix with coordinates of artificial reefs.}

\item{starting_values}{List with all starting value.}

\item{parameters}{List with all parameters.}

\item{random}{Numeric to randomize input values by 0 = 0 percent to 1 = 100 percent.}

\item{use_log}{Logical if TRUE, random log distribution is used}

\item{verbose}{If TRUE, progress reports are printed.}

\item{...}{Additional arguments passed on to \code{\link{raster}}.}
}
\value{
meta_sys
}
\description{
Initiate fish population.
}
\details{
Function to setup the fish population. If no fish shoud be created, set
\code{starting_values$pop_n = 0}.
}
\examples{
# Add example code

}
