% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nutr_input.R
\name{sim_nutr_input}
\alias{sim_nutr_input}
\title{sim_nutr_input}
\usage{
sim_nutr_input(
  n,
  max_i,
  freq_mn,
  input_max,
  variability,
  method = "noise",
  n_noise = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{freq_mn}{Numeric number of peak.}

\item{input_max}{Numeric with maximum input amount.}

\item{variability}{Variability of nutrient input}

\item{method}{String to specify how variability is simulated ('noise' or 'sd').}

\item{n_noise}{Integer with number of sine curves used for noise.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
vector
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curve. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_max} argument quantifies the
maximum value of the input "peaks".
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, freq_mn = 3,
input_max = 1, variability = 0.5)

}
