% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nutr_input.R
\name{sim_nutr_input}
\alias{sim_nutr_input}
\title{sim_nutr_input}
\usage{
sim_nutr_input(
  n,
  max_i,
  input_mn,
  freq_mn,
  variability = c(0, 0),
  method = "sd",
  n_noise = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{input_mn, freq_mn}{Numeric with mean input amount and frequency.}

\item{variability}{Variability of nutrient input.}

\item{method}{String to specify how variability is simulated ('noise' or 'sd').}

\item{n_noise}{Integer with number of sine curves used for noise.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
nutr_input
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curves. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_mn} argument quantifies the
maximum value of the input "peaks". The default \code{method = 'noise'} uses the mean
of \code{n_noise} sine curves to generate a noise signal.

If two \code{variability} parameters are provided, the first one is used for \code{input_mn},
the second one for \code{freq_mn}.

If \code{n_noise = NULL} (default), the value will be set to \code{n_noise = 5} automatically.
}
\examples{
nutr_input <- sim_nutr_input(n = 3, max_i = 4380, input_mn = 1, freq_mn = 3,
variability = 0.5)

}
