% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_input_sine.R
\name{simulate_input_sine}
\alias{simulate_input_sine}
\title{simulate_input_sine}
\usage{
simulate_input_sine(
  n,
  max_i,
  freq_mn,
  freq_sd,
  input_max,
  input_sd,
  phase = 0,
  return_df = FALSE
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{freq_mn, freq_sd}{Numeric number of peak.}

\item{input_max, input_sd}{Numeric with maximum input amount.}

\item{phase}{Numeric with sine curve parameters.}

\item{return_df}{Logical if TRUE data.frame is returned.}
}
\value{
vector
}
\description{
Initiate fish population.
}
\details{
Simulating nutrient input based on sine curve. The \code{freq_mn} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_max} argument quantifies the
maximum value of the input "peaks". Both arguments can be multiplied by the \code{*_sd}
argument and used as standard deviation of a normal distribution to add variability across
metaecosystems.
}
\examples{
# Add example code

}
