% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_input_sine.R
\name{simulate_input_sine}
\alias{simulate_input_sine}
\title{simulate_input_sine}
\usage{
simulate_input_sine(
  n,
  timestep,
  period_mn,
  period_sd,
  amplitude_mn,
  amplitude_sd,
  phase = 0,
  return_df = FALSE
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{timestep}{Vector timesteps}

\item{period_mn, period_sd}{Numeric with sine curve parameters.}

\item{amplitude_mn, amplitude_sd}{Numeric variability of parameters.}

\item{phase}{Numeric with sine curve parameters.}

\item{return_df}{Logical if data.frame is returned.}
}
\value{
vector
}
\description{
Initiate fish population.
}
\details{
ADD TEXT
}
\examples{
# Add example code

}
