% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement_meta.R
\name{simulate_movement_meta}
\alias{simulate_movement_meta}
\title{simulate_movement_meta}
\usage{
simulate_movement_meta(fishpop_values, n, pop_n, parameters, extent)
}
\arguments{
\item{fishpop_values}{List with fishpop matrices.}

\item{n}{Numeric with number of metaecosystems.}

\item{pop_n}{Numeric with initial number of fish individuals.}

\item{parameters}{List with all model parameters.}

\item{extent}{Spatial extent of the seafloor raster.}
}
\value{
vector
}
\description{
Initiate fish population.
}
\details{
ADD TEXT
}
\examples{
# Add example code

}
