% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nutr_input.R
\name{simulate_nutr_input}
\alias{simulate_nutr_input}
\title{simulate_nutr_input}
\usage{
simulate_nutr_input(
  n,
  max_i,
  frequency = 0,
  input_mn = 0,
  noise_sd = 0,
  amplitude_mn = 0,
  phase_mn = 0,
  amplitude_sd = 0,
  phase_sd = 0,
  verbose = TRUE
)
}
\arguments{
\item{n}{Integer with number of metaecosystems to setup.}

\item{max_i}{Integer with maximum number of simulation time steps.}

\item{frequency}{Numeric frequency.}

\item{input_mn}{Numeric with mean input amount and frequency.}

\item{noise_sd}{Numeric with noise added to sine functions.}

\item{amplitude_mn, phase_mn}{Numeric with mean amplitude and phase.}

\item{amplitude_sd, phase_sd}{Numeric with sd amplitude and phase.}

\item{verbose}{If TRUE, progress reports are printed.}
}
\value{
nutr_input
}
\description{
Simulate nutrient input.
}
\details{
Simulating nutrient input based on sine curves. The \code{frequency} argument quantifies
how many complete cycles of the sine function are present for a given \code{max_i},
i.e., how many "peaks" are present. The \code{input_mn} argument quantifies the
mean input values. Ampliude and phase mean and sd can be controlled by the corresponding
arguments.
}
\examples{
nutrients_input <- simulate_nutr_input(n = 3, max_i = 4380, frequency = 5, input_mn = 1)

}
