% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_meta.R
\name{summarize_meta}
\alias{summarize_meta}
\title{summarize_meta}
\usage{
summarize_meta(
  result,
  what = c("biomass", "production"),
  fun = "sum",
  na.rm = TRUE,
  lag = FALSE,
  return_df = TRUE
)
}
\arguments{
\item{result}{\code{meta_rn} object simulated \code{run_meta}.}

\item{what}{Logical if TRUE, the difference to the previous timestep is returned.}

\item{fun}{Function to aggregate results. Passed on to \code{aggregate}.}

\item{na.rm}{Logical passed on to \code{aggregate}.}

\item{lag}{Logical if TRUE, the difference to the previous timestep is returned.}

\item{return_df}{Logical if TRUE, a data.frame is returned.}
}
\value{
list
}
\description{
Summarize biomass and production values of each timestep
}
\details{
Summarize the biomass and production values for each local metaecosystem and timestep.
}
\examples{
\dontrun{
summarize_meta(result = result_rand)
}

}
